minimega: split mesh commands into _cli file
Follow pattern from most of minimega.
